requirements.txt
django==3.1.2
pymysql==0.10.1
djangorestframework==3.12.1
requests==2.24.0
